requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.9